refactor: dto, mapper를 뷰 계층으로 이동
- controller에 두면 추후 뷰가 변경됐을 때 해당 형식에 의존해야 하는 문제 발생
- dto, mapper는 현재 요구사항상 뷰의 데이터를 표현하기 위해 존재.
